Fixing MSVC warning in bam_stats.cpp.
git-svn-id: http://svn.seqan.de/seqan/trunk@11174 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/extras/demos/bam_stats.cpp
@@ -144,7 +144,7 @@ struct Stats
     String<unsigned> mismatchHisto;
     String<unsigned> insertHisto;
     String<unsigned> deletionHisto;
-    String<unsigned> avrgQuality;
+    String<double> avrgQuality;
 
     Stats() : numRecords(0), alignedRecords(0)
     {}
